requirements.txt
pytest==7.2.2
selenium==4.8.2
webdriver_manager==3.8.5
pytest_html==3.2.0
allure_pytest==2.13.0
requests==2.28.2